requirements.txt
astroid==2.3.3
bcrypt==3.1.7
blinker==1.4
cffi==1.14.0
click==7.1.1
cryptography==2.8
Flask==1.1.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
passlib==1.7.2
pkg-resources==0.0.0
psycopg2-binary==2.8.4
py-bcrypt==0.4
pycparser==2.20
pycrypto==2.6.1
pygal==2.4.0
pylint==2.4.4
six==1.14.0
SQLAlchemy==1.3.15
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
WTForms==2.2.1
